Introduce the Decimal object, a wrapper for BigDecimal, that provides a consistent decimal object. All the methods enforce the best practices recommended in BigDecimal.  It offers a simple API with no loss of precision. It's been painstakingly tested, although much work is still needed to make sure that it complies with the expected BigDecimal behaviour (when the latter is used correctly). All existing functionality in the statistical classes has been ported to use the new Decimal object.

+++ src/main/java/render/quantifyit/model/Probability.java
@@ -1,21 +1,24 @@
 package render.quantifyit.model;
 
+/**
+ * The probability of an event and the associated value
+ */
 public class Probability {
 	
-	private final double weight;
+	private final Decimal weight;
 	
-	private final double value;
+	private final Decimal value;
 
-	public Probability(final double weight, final double value){
+	public Probability(final Decimal weight, final Decimal value){
 		this.weight = weight;
 		this.value = value;
 	}
-	
-	public double getWeight() {
+
+	public Decimal getWeight() {
 		return weight;
 	}
 
-	public double getValue() {
+	public Decimal getValue() {
 		return value;
 	}
